pyproject.toml
[tool.poetry]
name = "contact-api"
version = "0.1.0"
description = ""
authors = ["Amolit <MikeKukPer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.111.0"}
sqlalchemy = "^2.0.30"
psycopg2 = "^2.9.9"
pydantic = "^2.7.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
alembic = "^1.13.1"
libgravatar = "^1.0.4"
asyncpg = "^0.29.0"
fastapi-mail = "^1.4.1"
jinja2 = "^3.1.4"
fastapi-limiter = "^0.1.6"
redis = "^5.0.4"
python-dotenv = "^1.0.1"
cloudinary = "^1.40.0"
pydantic-settings = "^2.2.1"
environs = "^11.0.0"
sphinx = "^7.3.7"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
